Prepared Fitindex data for entry in MySQL
I made the fitindexDataLoader and its respective tests to imput the data into MySQL based on the day it was collected. If there are multiple datapoints for the same day they are averaged.
I also moved SQLConnect.py into the base folder so it can be used by other programs.
Lastly I modified seperateData.py to make wrangling easier.